Show create group dialog when no groups exist
Introduces a CreateGroupDialog that prompts users to create a new group if none exist when adding a contact to a group. The dialog integrates with NewGroupChatSheet, allowing pre-selection of the contact to be added. Updates NewGroupChatSheet to support pre-selected contacts and refactors related logic in add_to_group_screen.dart.